requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.2
Flask==2.1.1
Flask-MySQLdb==1.0.1
idna==3.4
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
mariadb==1.0.10
MarkupSafe==2.1.1
mysqlclient==2.1.0
requests==2.28.2
urllib3==1.26.15
Werkzeug==2.1.1
zipp==3.8.0
